+++ packages/app/src/systems/Assets/hooks/useAsset.tsx
@@ -16,7 +16,7 @@ export const useAsset = (params?: {
 
   const asset = useMemo((): Asset | undefined => {
     // consider ETH as default asset
-    if (!ethTokenId && !fuelTokenId && !fuelContractId) {
+    if (!ethTokenId && !fuelTokenId) {
       return assets.find((asset) => asset.symbol === 'ETH');
     }
 
@@ -26,7 +26,6 @@ export const useAsset = (params?: {
 
       return (
         isSameEthAddress(ethAsset?.address, ethTokenId) ||
-        (fuelContractId && fuelAsset?.contractId === fuelContractId) ||
         (fuelTokenId && fuelAsset?.assetId === fuelTokenId)
       );
     });
